Change login and registration interface. Refactor.
