+++ backend/EGifts/EGifts/Messages/MessageNames/LoginNames.cs
@@ -1,0 +1,20 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Threading.Tasks;
+
+namespace EGifts.Messages.MessageNames
+{
+    public class LoginNames
+    {
+        public const string Result = "result";
+        public const string ResultMessage = "result_message";
+        public const string Login = "login";
+        public const string Mail = "mail";
+        public const string FirstName = "first_name";
+        public const string LastName = "last_name";
+        public const string Token = "token";
+        public const string BirthDate = "birth_date";
+        public const string Password = "password";
+    }
+}
