+++ backend/EGifts/EGifts/Program.cs
@@ -3,6 +3,7 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
 using EGifts.DataBase;
+using EGifts.Handlers;
 using EGifts.Messages;
 using Microsoft.AspNetCore.Hosting;
 using Microsoft.EntityFrameworkCore;
